+++ opyprint/__init__.py
@@ -1,4 +1,5 @@
 from .apply_style import apply_style
+from .format import format
 from .logger import Logger, PrintLogger, VoidLogger
 from .pp_context import PPContext
 from .pp_styles import PPStyles
@@ -9,19 +10,20 @@ from .utils import (
 )
 
 __all__ = [
-    'apply_style',
-    'Logger',
-    'PrintLogger',
-    'VoidLogger',
-    'PPContext',
-    'PPStyles',
-    'StyleOptions',
-    'dict_lt',
-    'is_dict',
-    'is_multiliner',
-    'is_oneliner',
-    'is_set',
-    'is_tuple',
-    'lt',
-    'print',
+    "apply_style",
+    "dict_lt",
+    "format",
+    "is_dict",
+    "is_multiliner",
+    "is_oneliner",
+    "is_set",
+    "is_tuple",
+    "Logger",
+    "lt",
+    "PPContext",
+    "PPStyles",
+    "PrintLogger",
+    "print",
+    "StyleOptions",
+    "VoidLogger",
 ]